Add JS helper to dynamically update favicon
Browsers like Chrome require that a favicon link be in the head to work properly.

Blazor does not allow data binding in the head tag, as this is considered static content.

By adding a JS helper to update the favicon at runtime, all browsers are compatible with the dynamic favicon. Any head content that needs to be dynamic should be handled in this way to maintain browser compatibilities.