Add BlazingUI and example
Kaiju under jorson pages contains an example of 3 buttons in a horizontal container with click and hover behaviors.

The BlazingUi library leaves the rendering calls and logic to the user.

Optimizations on re-rendering can be done by the user since they have hooks on each element for rendering.
If the user does not care to optimize rendering, they can simply re-render the whole tree each frame.