Convert to Blazor server project on DigitalOcean
P5 does not function properly due to a the asynchronous execution of
callback functions in the runtime. Since all JS interaction must be
asynchronous in Blazor server projects for now, the registered callbacks
for P5 run on a thread pool and cannot be synchronized.

Drawing logic for games is broken when done in C#, but P5 games can
still be drawn by writing the games in JS.